docs: add website/github badges
Because the configuration documentation tables get squashed in the GitHub UI, I've enabled GitHub Pages publishing of the README to `orbitalsync.com`. However, if one were to visit that page first it's not immediately clear how to access the actual source code/project/repo/issues/etc. By adding the badges to the top, users on GitHub can access the much easier to read 'published' version of the readme and visitors to the website can access the GitHub repo more easily.