feat: add distroless image
Distroless images are also available with `*-distroless` tags; for example, v1 is available as
`mattwebbio/orbital-sync:1-distroless`. These images are smaller and more secure than the
default Alpine-based images, because they contain only the Orbital Sync code and its direct
dependencies. They do not include a shell, package manager, or other tools that are typically
present in a Linux distribution. These will be the default images in a future major version
release.