fix: error "Not authorized" when uploading archives
The `persistentlogin=off` snippet in the request makes the server
to respond the cookie `persistentlogin`, which will then be send in the
POST `teleporter.php` request, which will cause the pihole server to
respond: "Not authorized" when this cookie is sent.

Removing this snippets from the login request removes the
`persistentlogin` from the response. Conesquently, the archive upload
is accepted by the pihole sever.